Backend Python API/requirements.txt
pydantic[email]==2.9.2
fastapi
uvicorn
langchain
langchain-groq
pydantic
python-multipart
python-dotenv
langgraph
langgraph-cli[inmem]
pytesseract
pypdfium2
Pillow
opencv-python
pandas
openpyxl
tabulate
camelot-py[cv]

#Langgraph related dependencies
langgraph-cli[inmem]
langgraph-checkpoint-mysql[aiomysql]
pymysql
qdrant-client
python-docx
PyPDF2
sentence-transformers
numpy
openai
langchain-openai